fix: #51876 修改int为float

+++ app/Resources/static-src/app/js/subject/index.js
@@ -270,7 +270,7 @@ export default class sbList {
     });
 
     $.validator.addMethod( 'noMoreThan', function(value, element, param) {
-      return parseFloat(value) <= parseInt($(param).val());
+      return parseFloat(value) <= parseFloat($(param).val());
     }, 'Please enter a lesser value.' );
   }
